extract common part of GithubPO.py and GooglePO.py to DefaultPO.py
