+++ src/po/GithubPO.py
@@ -1,8 +1,5 @@
-class GithubPO:
+import src.po.DefaultPO as dpo
 
-    def __init__(self, driver) -> None:
-        super().__init__()
-        self.driver = driver
 
-    def get_page_source(self):
-        return self.driver.page_source
+class GithubPO(dpo.DefaultPO):
+    pass
